refactoring of database java class
put all classes using JDBC to package database
