+++ src/java/control/AddItem.java
@@ -14,7 +14,7 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import model.ProductMgr;
+import database.ProductMgr;
 
 /**
  *
