+++ src/java/control/ViewItem.java
@@ -12,7 +12,7 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import model.ProductMgr;
+import database.ProductMgr;
 import model.Product;
 
 /**
