+++ src/java/database/BillMgr.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package model;
+package database;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -11,6 +11,11 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Calendar;
+import model.Bill;
+import model.BillAddress;
+import model.BillInfo;
+import model.Item;
+import model.User;
 
 /**
  *
