+++ src/java/database/FileMgr.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package model;
+package database;
 
 import java.io.File;
 import java.util.Iterator;
