+++ src/java/test/ProductMgrTest.java
@@ -86,7 +86,7 @@ public class ProductMgrTest {
     //getItemByProductId test
     
     ArrayList<Item> item = new ArrayList();
-    item = model.ProductMgr.getItemByProductId(222);
+    item = database.ProductMgr.getItemByProductId(222);
     System.out.println(item.size());
